Organize directory structure. Continue address code analysis comments
