+++ koala/quiz-evaluators/types.ts
