+++ koala/routers/manually-grade.ts
@@ -1,8 +1,8 @@
-import { errorReport } from "@/utils/error-report";
+import { errorReport } from "@/koala/error-report";
 import { z } from "zod";
 import { getUserSettings } from "../auth-helpers";
 import { prismaClient } from "../prisma-client";
-import { procedure } from "../trpc";
+import { procedure } from "../trpc-procedure";
 import { setGrade } from "./import-cards";
 
 export const manuallyGrade = procedure
