+++ koala/study_reducer.ts
