+++ pages/cards/[card_id].tsx
@@ -1,4 +1,4 @@
-import { trpc } from "@/utils/trpc";
+import { trpc } from "@/koala/trpc-config";
 import { Button, Checkbox, Container, Paper, TextInput } from "@mantine/core";
 import { useForm } from "@mantine/form";
 import { useRouter } from "next/router";
